Message model sets fields by itself
Many of the fields within the model are self-derived by the initial load
from Outlook. So instead of having to manually set these fields prior to
saving the object somewhere, udpated the object to be smart enough to
set those properties itself.

Ensure we don't change the way it serialises or deserialises by ignore the private properties