BUG: Resolve duplicate objects inserted into EF tracking
The FInd() method retrieves the row from the DB, creates and object and
starts tracking it. If we then try and put our object (with the same ID)
into the changeset (the Update() line), then it notices the keys collide
and throws an error about have keys with the same ID.

To avoid this, use AsNoTracking to do a read-only query that doesn't set
up object tracking, and avoid using Find (which doesn't support No
Tracking).